fix: Aplicadas correções críticas de Code Review (PR#7)
