+++ app/app/Libraries/Pagamentos/Entities/Cliente.php
