+++ app/app/Libraries/Pagamentos/Entities/Transacao.php
