+++ app/app/Libraries/Pagamentos/Entities/_EntityBase.php
