+++ app/composer.json
@@ -11,8 +11,9 @@
     },
     "require": {
         "php": "^8.1",
-        "ariellcannal/pagamentos": "@dev",
-        "codeigniter4/framework": "^4.0"
+        
+        "codeigniter4/framework": "^4.0",
+        "guzzlehttp/guzzle": "^7.0"
     },
     "require-dev": {
         "fakerphp/faker": "^1.9",
@@ -40,11 +41,5 @@
     },
     "scripts": {
         "test": "phpunit"
-    },
-    "repositories": {
-        "pagamentos": {
-            "type": "path",
-            "url": "../"
-        }
     }
 }
