--- composer.json
@@ -1,18 +1,0 @@
-{
-    "description": "CANNAL - Pagamentos",
-    "name": "ariellcannal/pagamentos",
-    "type": "project",
-    "license": "MIT",
-    "require": {
-		"php": ">=8.1",
-		"psr/log": "^3.0",
-        "pagarme/pagarme-php-sdk": "~6.8"
-    },
-    "autoload": {
-        "psr-4": {
-            "CANNALPagamentos\\": "src/"
-        }
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true
-}
